** REPLACED apply_natural_column_order() WITH TOTALLY NEW LOGIC
The apply_natural_column_order() method gets called to put columns back
into an order in line with the actual database schema, but there are a
number of complications and details which I do not fully remember
anymore, or exactly why this was needed in the first place. It may just
bea hack/bandaid over some deeper problem that was never resolved (at
the time, a major refactor was planned which never ended up happening).
What I do know is that the algorithm was very slow and this becomes a
huge problem in cases where there are MANY columns on a single
table/view.

The old logic called through the "colspec" guts which is the technically
most correct code path, but just tooo slow.

I have rewritten the logic from scratch to just walk the DBIC
relationships and put the orders back to match the DBIC schema. This is
massively faster, but it does not produce exactly the same order as the
previous logic. I'm not sure if this is a problem or not. The new,
faster code may even be more correct, and the differences may not come
up in any meaningful way except in extreme edge cases. The main thing I
am concerned about is the logic whic moves the location of single rels
next to their associated FK, when teh rel name and FK name are not the
same.

More testing and investigation needs to be done, so I've left the old
code in the file with an easy way to flip back and forth between new and
old. But, as of this commit, it is using the new logic, to force
failure, since that is the only way to figure it out at this point.

This is probably TO BE CONTINUED...