"canary" test case for internals of QuickSearch
This test case verifies the ability to override pieces of quicksearch.
There isn't an official API for this yet, but there is at least one
project out there using this method, which will need updated if the
internals of quicksearch ever change.

This example overrides chain_Rs_req_quicksearch to make a change to the
returned resultset.  (doesn't matter what, just that the code ran when
overriding that method)

Then it also overrides _resolve_quicksearch_condition to make one
per-column change to the way that the search expression is generated.
This couldn't be done in chain_Rs_req_quicksearch because all the
columns together get applied in one single { -or => [] } clause.
(subclassing chain_Rs_req_quicksearch would be sufficient if there were
 an official API to reach back into the ResultSet clauses and inspect
 and modify them)