WIP - PRELIMINARY BACKEND Quick Search -> SearchBox REFACTOR
Refactoring of the back-end code of grid quick search to use a new
object class API called 'SearchBox'. The purpose of this is to provide
an easy way to hook into and customize the Quick Search box. So far,
only back-end code has been changed and reproduces current, simple
"like" and "exact" mode operation. For this feature to have any benefit,
and for additional SearchBox classes to be able tom be addded, extensive
refactoring of front-end JavaScript code must also happen (which is what
I'm doing next)

API CHANGE: _resolve_quicksearch_condition method no longer called. This
was a private method, however, so it should not have been munged by any
end-user apps (but, if any apps in the wild did anyway, they will break
with this commit)

IN PROGRESS...

(set dev version 1.3202_01)