SearchBox: NOW DRIVES FRONT-END QUICK SEARCH UI
The front-end grid quick search UI is now controlled by the newly added
'searchbox_modes' DbicLnk API/config. As a proof-of-concept of this, a
3rd quick search/search box mode has been added called "AnyKeywords".
This will probably not be left in place as-os. The main purpose of this
new class was as a test.

The new "keyword" quick search mode added in this commit, although
created as a test, is useful, as it splits on whitespace and matches any
of the words rather than the entire string.

The changes to the front-end code are very minimal. It still functions
the same way, it just builds the menu according to the new
'searchbox_configs' ext config instead of the previous hard-coded
"Normal" or "Exact" and now the front and back ends are tied together
ina  very clean, but very powerful, object class API.

This commit also has a number of back-end code improvements refining the
API from the previous initial commit.

This is still in progress; not all features have been implemented yet,
such as the planned 'choose_columns' setting, so the API is still
changing and no production apps should be ran on this commit.

IN PROGRESS...

(set dev version 1.3202_02)