SearchBox: ADD A COMPREHENSIVE DOCUMENTATION FEATURE
With the new ability to have ultra customized behavioiurs of varied
search box modes, it became obvious that there was have to be a way to
communicate to the end user what each mode does. This has been added to
the API, with a required 'documentation' attr, and a new menu point
which displays the mode documentation in a clear and concice manner.