Grid QUICK SEARCH value saved in saved searches
By default, the value of the quick search box is now included in the
saved search data, so that when a search is saved that has a value
stored in the search box, that value is preloaded when the search is
loaded.

This is an API change, but the old behaviour can be turned back on by
setting the Grid module attr quicksearch_value_in_saved_search to '0'
(false).

The reason that this has been made the default is that the Quick Search
is now a more rubust and customizable part of the view, with the new
SearchBox modes. It makes sense now that this would be the default
behaviour.

(set dev version 1.3202_04)