Expose the SearchBox per-column condition building method
The default implementation of SearchBox calculates a condition
for each searched column and then "OR"s them together.  The
implementation of these conditions is complex and relies on
private methods.  In order for subclasses to be able to alter
the implementation without completely rewriting it, they need
access to call the per-column method.  This change refactors
the _get_query_condition_list so that the per-column portion
is public, as "get_condition_for_column".

I also added documentation for the API of chain_query_search_rs
to both SearchBox and SearchBox::Normal