Ensure joins using where cluse instead of select - #189
This is an alternate fix for the previous solution, where I
add the dummy column in the where clause rather than needing
to select it.  This avoids a potential (though unlikely)
problem where the column chosen might be expensive to fetch
and lead to a needless performance hit.  By making it into
a 'col = col' expression, hopefully the database engine
optimizes it out.