Use "column IS NULL or column IS NOT NULL" instead of "column=column" - #189
I previously chose "column = column" as a no-op true expression to
force DBIC to include a join.  This does not work in all cases because
a NULL field is never equal to a NULL field, per SQL rules.  The result
is that rows go missing from the grid when this code is triggered.

An actual fix is to say the column may be null or not null.