SWITCH TO encode_json_ascii (FROM *_utf8) ACROSS BOARD + TPL ENC - fixes #192
Switched from using encode_json_utf8() to encode_json_ascii()
across-the-board.

encode_json_ascii, not encode_json_utf8, is the actual safe best way to
encode content which may include utf-8 characters, because the "ascii"
part is itself a safe encoding which *is* still aware of UTF-8, it just
encodes the UTF-8 in ASCII with escape sequences, which compliant JSON
decoders will turn back into utf-8 at the other end. This is just a
feature of JSON that we haven't been using but should, since this saves
us from having to worry about properly handling utf8 outside the JSON
encapsulation, and removes a whole layer of text processing where
unicode could go wrong.

Also removed a very old extra UTF-8 encoding in the template view path,
which was a workaround (just based on the old comment) for something
that is likely long gone/no longer needed, and this further could be a
place where unicode things could go wrong. I didn't directly observe the
problem described in #192, but agree this code shouldn't be there
regardless, so I removed it, but left it commented out should the actual
workaround this was for still be out there and this change suddenly
breaks it. Should that happen, the real/correct solution must be found