DbicPropPage: skip multi-rel grids for unsupported coderef conds
The DbicPropPage module displays property grids for every included
single rel column in the include_colspec (more common) and multi/row
grids for every multi relationship column in the include_colspec (less
common). The way the latter is implemented cannot support relationships
if the cond is defined by a coderef, and when it tries it barfs, so
these are now just excluded/skipped, which probably won’t even be
noticed in most cases. This is a very edge-case situation and the
implementation of this is wholly contained and specific to just the
DbicPropPage and not anyplace else in RapidApp. If this feature is ever
important enough to work for these cases, it would require a whole new
implementation, because of the front-side design, it can never work with
coderefs because these must be dealt with in perl code in the backend