New MdPopupEditor allows windowed Markdown editing
The existing MdEditor does not place nicely with the default
editing contexts of grids and attribute views, because editors
normally receive a small editing boundary, and because it
doesn't cleanly handle focus changes to save the edits.

This new edit component can be used instead, and opens a
pop-up window that shows the MdEditor along with OK/Cancel
buttons to commit or abort the edit.