Replace Hash::Merge with a faster custom implementation
The original Hash::Merge has a poor API that allows people to
change the global default behavior.  RapidApp made the assumption
that it was always RIGHT_PRECEDENT.  Later, a copy of Hash::Merge
was included directly in RapidApp, which avoided that problem,
but the implementation still had a lot of flexibility that wasn't
being used.  By rewriting it into a single sub that only calls
one function per depth of hash, and one call to clone(), it is
dramatically faster, and knocks out several percent of startup
delay.