Dramatic speedup of colspec_select_columns
This function had received a lot of caching triage previously,
but I was able to rearrange the algorithm itself for so much speed
gain that the caching was no longer helpful.  So, now we both get
a speedup and get to drop the complexity of the caching.

I also moved the function out of the Role into the Util module so
that we can experiment with an XS version, if we want.