Experimental ThreadSocketWatch using Mike's Inline C thread
Experimental wrapping of all requests with a thread which watches the
client psgi request socket and detects if it is closed/aborted. This is
using Inline::C code written by @nrdvana but with my implementation.
Unfortunately, cancelling a request causes the whole server to seg fault
currently (using plackup -s Twiggy -p 5000)