NOT WORKING MiD-DEBUG CODE - DO NOR USE
Saving that lateb state of Mike and my experiemnting with ideas when we
ran into the brutal fact that MySQL c lib ignores signals and cannot be
interrupted. this blows up our whole theory for being able to abort
queries.

Mike tried some crazy ideas, closing the mysql socket from his C thread
code, but this is probably unsafe and unworkable.

Committing this code in case it has vlue later. We have to give up on
this for now, but we made AMAZING progress and have a fully working
implementstion which is able to detech client abort and throw an
excepotion,. This opens the door to AMNY possibilities. But, DBD::mysql
queries ignore signals and so we still cant interrupt them using this
method.