New debugging tool, render_fd_table
This dumps out a list of all the file descriptors, using proc to
find out what files are open, and getpeeraddr to find out the
remote IP address and port of the connections.

This should help discover how libmysqlclient is able to continue
the request after we close the socket we think it is using.