Rework ExcelExport and ExcelTableWriter to support column types
The original ExcelTableWriter supported distinction between string
and non-string columns, but the ExcelExport role never set this
flag, and instead just let the XLSX module auto-detect all values.
This makes it impossible to export IDs that have leading zeroes
and re-import that worksheet later.

The fix was more extensive than I expected - I realized that there
was also a problem with string values beginning with '=' to be
treated as formulas, and result in execution of arbitrary code
depending on Excel configuration.

To fully solve the problem, I added a more detailed type to each
ExcelTableWriter column definition, and use that to call the
appropriate 'write' method on the worksheet.

(code not tested yet)